+++ src/main/java/step1/domain/Operand.java
@@ -1,4 +1,4 @@
-package step1;
+package step1.domain;
 
 import java.util.Objects;
