+++ src/test/java/step1/domain/ExpressionTest.java
@@ -1,4 +1,4 @@
-package step1;
+package step1.domain;
 
 
 import org.junit.jupiter.api.Test;
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Queue;
 
 import static org.assertj.core.api.Assertions.*;
-import static step1.Operator.*;
+import static step1.domain.Operator.*;
 
 public class ExpressionTest {
