+++ src/test/java/step1/domain/OperatorTest.java
@@ -1,4 +1,4 @@
-package step1;
+package step1.domain;
 
 import org.junit.jupiter.api.Test;
